review comments, more tests
