+++ servicetalk-concurrent-reactivestreams/src/test/java/io/servicetalk/concurrent/reactivestreams/tck/PublisherConcatWithCompletableCancelTckTest.java
@@ -24,6 +24,6 @@ import org.testng.annotations.Test;
 public class PublisherConcatWithCompletableCancelTckTest extends AbstractPublisherOperatorTckTest<Integer> {
     @Override
     protected Publisher<Integer> composePublisher(Publisher<Integer> publisher, int elements) {
-        return publisher.concat(Completable.completed(), true);
+        return publisher.concatPropagateCancel(Completable.completed());
     }
 }
